[feat] 약관, 회원가입 페이지 통합
