+++ src/app/(non-auth)/signup/page.tsx
@@ -1,84 +1,93 @@
 'use client';
-
-import Link from 'next/link';
-import { assetsStorageUrl } from '@/utils/supabase/storage';
-import { useState } from 'react';
-import { useAuth } from '@/hooks/useAuth';
-import { toast } from 'react-toastify';
+import Terms from '@/constant/Terms';
+import PrivacyPolicy from '@/constant/PrivacyPolicy';
+import { useState, useEffect } from 'react';
 import { useRouter } from 'next/navigation';
-import Image from 'next/image';
+import { toast } from 'react-toastify';
+import { useAuth } from '@/hooks/useAuth';
 
-const SignUpPage = () => {
+const TermsPage = () => {
+  const [allChecked, setAllChecked] = useState(false);
+  const [termsChecked, setTermsChecked] = useState(false);
+  const [privacyChecked, setPrivacyChecked] = useState(false);
   const [email, setEmail] = useState('');
   const [password, setPassword] = useState('');
-  const [showPasswordImage, setShowPasswordImage] = useState(false);
   const [confirmPassword, setConfirmPassword] = useState('');
   const { signUp, error } = useAuth();
   const router = useRouter();
 
-  const handleSignUp = async (e: React.FormEvent) => {
+  useEffect(() => {
+    setTermsChecked(allChecked);
+    setPrivacyChecked(allChecked);
+  }, [allChecked]);
+
+  // 개별 체크박스 상태 변경 핸들러
+  const handleTermsChange = () => {
+    setTermsChecked(!termsChecked);
+  };
+
+  const handlePrivacyChange = () => {
+    setPrivacyChecked(!privacyChecked);
+  };
+
+  /** 모든 체크박스가 선택되었는지 확인 */
+  useEffect(() => {
+    setAllChecked(termsChecked && privacyChecked);
+  }, [termsChecked, privacyChecked]);
+
+  const handleSubmit = async (e: React.FormEvent<HTMLFormElement>) => {
     e.preventDefault();
+    
+    // 약관 동의를 확인합니다.
+    if (!termsChecked || !privacyChecked) {
+      toast('모든 약관에 동의해야 합니다.');
+      return;
+    }
+    
+    // 비밀번호 일치 여부를 확인합니다.
     if (password !== confirmPassword) {
       toast.error('비밀번호가 서로 일치하지 않습니다.');
       return;
     }
+    
+    // 비밀번호 길이를 확인합니다.
     if (password.length < 6) {
       toast.error('비밀번호는 최소 6자 이상이어야 합니다.');
       return;
     }
+    
+    // 회원가입을 시도하고 결과를 처리합니다.
     const signUpResult = await signUp(email, password);
-
+    
     if (signUpResult && signUpResult.error) {
       toast.error(signUpResult.errorMessage);
       return;
     }
-
+    
+    // 회원가입 성공 메시지를 표시하고 로그인 페이지로 이동합니다.
     toast.success('회원가입에 성공했습니다! 로그인 페이지로 이동합니다.');
     router.push('/login');
   };
+  
 
   const handlePasswordChange = (e: React.ChangeEvent<HTMLInputElement>) => {
     setPassword(e.target.value);
-    setShowPasswordImage(!!e.target.value);
   };
 
   const handleConfirmPasswordChange = (e: React.ChangeEvent<HTMLInputElement>) => {
     setConfirmPassword(e.target.value);
-    setShowPasswordImage(!!e.target.value);
   };
 
   return (
-    <article className="flex flex-col h-[84vh]">
-      <div className="grid grid-cols-10 min-h-full bg-bgColor1">
-        <div className="col-span-4 flex items-center justify-center border-r-2 border-solid border-pointColor1">
-          <div className="mr-4 w-[300px] h-[300px] bg-bgColor2 rounded-full flex items-center justify-center overflow-hidden border border-solid border-pointColor1">
-            {showPasswordImage ? (
-              <Image
-                src={`${assetsStorageUrl}/login_2.png`}
-                alt="로그인 이미지"
-                width={300}
-                height={300}
-                quality={100}
-                className="w-full h-full object-cover rounded-full"
-              />
-            ) : (
-              <Image
-                src={`${assetsStorageUrl}/login_1.png`}
-                alt="로그인 이미지"
-                width={300}
-                height={300}
-                quality={100}
-                className="w-full h-full object-cover rounded-full"
-              />
-            )}
-          </div>
-        </div>
-        <div className="col-span-6 flex items-center bg-white justify-center">
+    <article className="min-h-screen flex items-center justify-center">
+      <div className="flex w-[600px] items-center justify-center">
+        <div className="w-full bg-white p-3 rounded mx-4">
+          <h1 className="text-xl font-bold text-pointColor1 text-center mt-4 mb-6">회원 가입</h1>
+          <div className="col-span-6 flex items-center bg-white justify-center">
           <div className="w-full max-w-md p-8 bg-white">
-            <form className="space-y-6" onSubmit={handleSignUp}>
-              <div className="flex flex-col gap-2">
+              <div className="flex flex-col gap-4">
                 <div>
-                  <label htmlFor="email" className="text-sm font-medium text-gray-700"></label>
+                  <label htmlFor="email" className="text-md font-bold text-blackColor">이메일</label>
                   <div className="mt-1">
                     <input
                       id="email"
@@ -93,7 +102,7 @@ const SignUpPage = () => {
                   </div>
                 </div>
                 <div>
-                  <label htmlFor="password" className="text-sm font-medium text-gray-700"></label>
+                  <label htmlFor="password" className="text-md font-bold text-blackColor">비밀번호</label>
                   <div className="mt-1">
                     <input
                       id="password"
@@ -108,7 +117,7 @@ const SignUpPage = () => {
                   </div>
                 </div>
                 <div>
-                  <label htmlFor="confirm-password" className="text-sm font-medium text-gray-700"></label>
+                  <label htmlFor="confirm-password" className="text-md font-bold text-blackColor">비밀번호 확인</label>
                   <div className="mt-1">
                     <input
                       id="confirm-password"
@@ -125,24 +134,92 @@ const SignUpPage = () => {
               </div>
               {error && <div className="text-red-500">{error}</div>}
               <div>
-                <button
-                  type="submit"
-                  className="w-full px-4 py-5 rounded-md text-white bg-pointColor1 shadow hover:bg-gray-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-offset-gray-100 focus:ring-black"
-                >
-                  회원가입
-                </button>
               </div>
-            </form>
-            <div className="flex justify-center mt-3">
-              <Link className="text-sm font-medium text-pointColor1 hover:underline" href="/login">
-                이미 계정이 있나요? 로그인하기
-              </Link>
-            </div>
           </div>
+        </div>
+          <form className="space-y-6 mt-4 mb-5" onSubmit={handleSubmit}>
+            <div className="flex items-start">
+            <label htmlFor="terms-checkbox" className="inline-flex items-center cursor-pointer">
+              <input
+                id="terms-checkbox"
+                type="checkbox"
+                className="sr-only" 
+                checked={termsChecked}
+                onChange={handleTermsChange}
+              />
+              <span className={`relative w-5 h-5 rounded-full border-solid border-2 ${termsChecked ? 'bg-pointColor1 border-pointColor1' : 'border-pointColor1'}`}>
+                {termsChecked && (
+                  <span className="absolute inset-0 flex items-center justify-center">
+                    <span className="w-2.5 h-2.5 bg-white rounded-full"></span>
+                  </span>
+                )}
+              </span>
+              <label htmlFor="privacy-checkbox" className="ml-2 text-md font-bold">
+              <span className="text-pointColor1 text-md font-bold">[필수]</span> 뭔말Easy? 이용약관
+                </label>
+            </label>
+            </div>
+            <div className="rounded-md border border-solid border-pointColor1 p-4 space-y-4 overflow-auto max-h-40">
+              <Terms />
+            </div>
+            <div className="flex items-start">
+            <label htmlFor="privacy-checkbox" className="inline-flex items-center cursor-pointer">
+              <input
+                id="privacy-checkbox"
+                type="checkbox"
+                className="sr-only" 
+                checked={privacyChecked}
+                onChange={handlePrivacyChange}
+              />
+              <span className={`relative w-5 h-5 rounded-full border-solid border-2 ${privacyChecked ? 'bg-pointColor1 border-pointColor1' : 'border-pointColor1'}`}>
+                {privacyChecked && (
+                  <span className="absolute inset-0 flex items-center justify-center">
+                    <span className="w-2.5 h-2.5 bg-white rounded-full"></span>
+                  </span>
+                )}
+              </span>
+              <label htmlFor="privacy-checkbox" className="ml-2 text-md font-bold">
+                <span className="text-pointColor1 text-md font-bold">[필수]</span> 개인정보 수집 및 이용
+              </label>
+            </label>
+            </div>
+            <div className="rounded-md border border-solid border-pointColor1 p-4 space-y-4 overflow-auto max-h-40">
+              <PrivacyPolicy />
+            </div>
+            <div className="flex items-start">
+              <label htmlFor="all-checkbox" className="inline-flex items-center cursor-pointer">
+              <input
+                id="all-checkbox"
+                type="checkbox"
+                className="sr-only" 
+                checked={allChecked}
+                onChange={() => setAllChecked(!allChecked)}
+              />
+              <span className={`relative w-5 h-5 rounded-full border-solid border-2 ${allChecked ? 'bg-pointColor1 border-pointColor1' : 'border-pointColor1'}`}>
+                {allChecked && (
+                  <span className="absolute inset-0 flex items-center justify-center">
+                    <span className="w-2.5 h-2.5 bg-white rounded-full"></span>
+                  </span>
+                )}
+              </span> 
+              <label htmlFor="all-checkbox" className="ml-2 text-md font-bold leading-tight">
+                전체 동의하기
+              </label>
+              </label>
+            </div>
+            <div className="flex">
+              <button
+                type="submit"
+                className="w-full px-6 py-2 border border-transparent rounded-md shadow-sm text-base font-bold text-white bg-pointColor1 hover:bg-blue-600 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-pointColor1"
+              >
+                가입 완료
+              </button>
+            </div>
+          </form>
         </div>
       </div>
     </article>
   );
 };
 
-export default SignUpPage;
+export default TermsPage;
